pyproject.toml
[project]
name = "phonemizer-api"
version = "0.1.0"
description = "API service for text to phoneme conversion"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "gruut>=2.0.0",
    "piper-phonemize-cross>=1.0.0",
    "pydantic>=1.8.2",
]
requires-python = ">=3.9,<3.11"

[project.optional-dependencies]
dev = [
    "pytest>=6.2.5",
    "httpx>=0.19.0",
    "pytest-asyncio>=0.15.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"] 